Closes #456 - CI/CD automated builds
Database::newInventory expects the TableSchema for the parent (eg, yeast
or misc) as the parameter.

I got it right in exactly half of the cases. In the other half of the
cases, I was sending the inventory tables TableSchema, which caused a
sigsegv later on.

My rule of thumb is to not burden methods with knowledge they don't
need. The four calls where I was sending the inventory table TableSchema
have been fixed.

All of the tests now pass on my machine.